# Conflicts:
#	Mimir.Shared/Services/HeadlessStateService.cs
#	Mimir.Shared/Services/IStateService.cs
#	Mimir.Worker/Constants/PlanetType.cs
#	Mimir.Worker/Handler/Balance/NcgBalanceHandler.cs

+++ Mimir.Worker/Services/IItemProductCalculationService.cs
@@ -1,10 +1,13 @@
+using Mimir.Shared.Constants;
+using Mimir.Shared.Client;
+using Mimir.Shared.Services;
 using Lib9c.Models.Market;
 
 namespace Mimir.Worker.Services;
 
 public interface IItemProductCalculationService
 {
-    Task<int?> CalculateCombatPointAsync(ItemProduct itemProduct);
+    Task<long?> CalculateCombatPointAsync(ItemProduct itemProduct);
 
     Task<(int? crystal, int? crystalPerPrice)> CalculateCrystalMetricsAsync(ItemProduct itemProduct);
 }
